fix stories and pages/ include require_api_key

+++ flask_api/app/routes/stories.py
@@ -43,7 +43,8 @@ def get_start_page(story_id):
 @bp.post("")
 def create_story():
     r = require_api_key()
-    if r: return r
+    if r:
+        return r
     data = request.get_json(force=True)
 
     story = Story(
@@ -59,6 +60,9 @@ def create_story():
 
 @bp.put("/<int:story_id>")
 def update_story(story_id):
+    r = require_api_key()
+    if r:
+        return r
     story = Story.query.get_or_404(story_id)
     data = request.get_json(force=True)
 
@@ -73,6 +77,9 @@ def update_story(story_id):
 
 @bp.delete("/<int:story_id>")
 def delete_story(story_id):
+    r = require_api_key()
+    if r:
+        return r
     story = Story.query.get_or_404(story_id)
 
     db.session.delete(story)
@@ -80,9 +87,12 @@ def delete_story(story_id):
     return "", 204
 
 
-
 @bp.post("/<int:story_id>/pages")
 def create_page(story_id):
+    r = require_api_key()
+    if r:
+        return r
+
     Story.query.get_or_404(story_id)
     data = request.get_json(force=True)
 
@@ -97,4 +107,3 @@ def create_page(story_id):
     db.session.commit()
 
     return jsonify(page.to_dict()), 201
-
